feat(chat/native): moved ChatAndPollsNavigator to mobile/navigation
