+++ react/features/chat/components/native/index.js
@@ -1,6 +1,6 @@
 // @flow
 
 export { default as Chat } from './Chat';
-export { default as ChatAndPolls } from './ChatAndPolls';
+export { default as ChatAndPolls } from '../../../mobile/navigation/components/chat/components/ChatAndPollsNavigator';
 export { default as ChatButton } from './ChatButton';
 export { default as ChatPrivacyDialog } from './ChatPrivacyDialog';
